Refactor: DDD 스러운 클래스 분리
DDD는 Root 도메인으로부터 하위 도메인을 다루기 때문에 PaymentDetailController부터 시작되는 구조로 변경
